# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kleines Mypage API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/profiles/{id}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update profile by user ID",
                "description": "Update profile by user ID. Only `MASTER` and `MANAGER` are allowed to update profiles.",
                "operationId": "643db8f121edcfb33ecefca44f08f8fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Profile object that needs to be updated",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/Profile"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "example": "Bad Request"
                                                },
                                                "status": {
                                                    "type": "integer",
                                                    "example": 400
                                                },
                                                "detail": {
                                                    "allOf": [
                                                        {
                                                            "properties": {
                                                                "part": {
                                                                    "type": "array",
                                                                    "items": {
                                                                        "type": "string",
                                                                        "example": "The selected part is invalid."
                                                                    }
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "example": "Unauthorized"
                                                },
                                                "status": {
                                                    "type": "integer",
                                                    "example": 401
                                                },
                                                "detail": {
                                                    "type": "string",
                                                    "example": "Unauthorized"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "example": "Forbidden"
                                                },
                                                "status": {
                                                    "type": "integer",
                                                    "example": 403
                                                },
                                                "detail": {
                                                    "type": "string",
                                                    "example": "Forbidden"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Profile": {
                "properties": {
                    "last_name": {
                        "type": "string",
                        "example": "山田"
                    },
                    "first_name": {
                        "type": "string",
                        "example": "太郎"
                    },
                    "name_kana": {
                        "type": "string",
                        "example": "ヤマダタロウ"
                    },
                    "grade": {
                        "type": "integer",
                        "example": 18
                    },
                    "part": {
                        "type": "string",
                        "example": "T"
                    }
                },
                "type": "object"
            },
...
        }
    }
}